Move the pull command to its own class
