+++ spec/commands/pull_spec.rb
@@ -1,0 +1,11 @@
+require 'spec_helper'
+
+describe 'pull' do
+
+  xit 'needs testing'
+
+  describe '--all' do
+    xit 'needs testing'
+  end
+
+end
